Fix FindAllTerraformModulePathsInRoot and move
