+++ modules/terraform/validate.go
@@ -1,41 +1,9 @@
 package terraform
 
 import (
-	// We alias Golang's native testing package to go_test to avoid naming conflicts with terratest's own testing module
-
-	"path/filepath"
-	go_test "testing"
-
-	"github.com/gruntwork-io/terratest/modules/files"
 	"github.com/gruntwork-io/terratest/modules/testing"
 	"github.com/stretchr/testify/require"
 )
-
-// ValidateAllTerraformModules automatically finds all folders specified in RootDir that contain .tf files and runs
-// InitAndValidate in all of them.
-// Filters down to only those paths passed in ValidationOptions.IncludeDirs, if passed.
-// Excludes any folders specified in the ValidationOptions.ExcludeDirs. IncludeDirs will take precedence over ExcludeDirs
-// Use the NewValidationOptions method to pass relative paths for either of these options to have the full paths built
-// Note that go_test is an alias to Golang's native testing package created to avoid naming conflicts with Terratest's
-// own testing package. We are using the native testing.T here because Terratest's testing.T struct does not implement Run
-func ValidateAllTerraformModules(t *go_test.T, opts *ValidationOptions) {
-	dirsToValidate, readErr := FindTerraformModulePathsInRootE(opts)
-	require.NoError(t, readErr)
-
-	for _, dir := range dirsToValidate {
-		dir := dir
-		t.Run(dir, func(t *go_test.T) {
-			t.Parallel()
-
-			//Copy the module to a tmp directory to avoid conflicts with tests that don't copy to /tmp
-			testFolder, err := files.CopyTerraformFolderToTemp(opts.RootDir, filepath.Base(dir))
-			require.NoError(t, err)
-
-			tfOpts := &Options{TerraformDir: testFolder}
-			InitAndValidate(t, tfOpts)
-		})
-	}
-}
 
 // Validate calls terraform validate and returns stdout/stderr.
 func Validate(t testing.TestingT, options *Options) string {
