+++ modules/test-structure/test_structure.go
@@ -7,8 +7,11 @@ import (
 	"path/filepath"
 	"strings"
 
+	go_test "testing"
+
 	"github.com/gruntwork-io/terratest/modules/files"
 	"github.com/gruntwork-io/terratest/modules/logger"
+	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/gruntwork-io/terratest/modules/testing"
 	"github.com/stretchr/testify/require"
 )
@@ -97,4 +100,40 @@ func CopyTerraformFolderToTemp(t testing.TestingT, rootFolder string, terraformM
 func cleanName(originalName string) string {
 	parts := strings.Split(originalName, "/")
 	return parts[len(parts)-1]
+}
+
+// ValidateAllTerraformModules automatically finds all folders specified in RootDir that contain .tf files and runs
+// InitAndValidate in all of them.
+// Filters down to only those paths passed in ValidationOptions.IncludeDirs, if passed.
+// Excludes any folders specified in the ValidationOptions.ExcludeDirs. IncludeDirs will take precedence over ExcludeDirs
+// Use the NewValidationOptions method to pass relative paths for either of these options to have the full paths built
+// Note that go_test is an alias to Golang's native testing package created to avoid naming conflicts with Terratest's
+// own testing package. We are using the native testing.T here because Terratest's testing.T struct does not implement Run
+func ValidateAllTerraformModules(t *go_test.T, opts *ValidationOptions) {
+	dirsToValidate, readErr := FindTerraformModulePathsInRootE(opts)
+	require.NoError(t, readErr)
+
+	for _, dir := range dirsToValidate {
+		dir := dir
+		t.Run(strings.TrimLeft(dir, "/"), func(t *go_test.T) {
+			t.Parallel()
+
+			// Determine the absolute path to the git repository root
+			cwd, cwdErr := os.Getwd()
+			require.NoError(t, cwdErr)
+			gitRoot, gitRootErr := filepath.Abs(filepath.Join(cwd, "../../"))
+			require.NoError(t, gitRootErr)
+
+			// Determine the relative path to the example, module, etc that is currently being considered
+			relativePath, pathErr := filepath.Rel(gitRoot, dir)
+			require.NoError(t, pathErr)
+			// Copy git root to tmp and supply the path to the current module to run init and validate on
+			testFolder := CopyTerraformFolderToTemp(t, gitRoot, relativePath)
+			require.NotNil(t, testFolder)
+			// Run Terraform init and terraform validate on the test folder that was copied to /tmp
+			// to avoid any potential conflicts with tests that may not use the same copy to /tmp behavior
+			tfOpts := &terraform.Options{TerraformDir: testFolder}
+			terraform.InitAndValidate(t, tfOpts)
+		})
+	}
 }
